fix: `s3_force_single_request` not parsed
